requirements/ci.txt
asgiref==3.3.1
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
boto==2.49.0
boto3==1.16.4
botocore==1.19.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.3.1
coveralls==3.0.0
cryptography==3.4.5
decorator==4.4.2
defusedxml==0.6.0
django-allauth==0.44.0
django-autocomplete-light==3.5.1
django-autoslug==1.9.8
django-braces==1.14.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.4.0
django-grappelli==2.14.3
django-model-utils==4.1.1
django-ratelimit==3.0.1
django-redis==4.12.1
django-reversion==3.0.9
django-storages==1.11.1
django-test-plus==1.4.0
django-webtest==1.9.7
docopt==0.6.2
factory-boy==3.2.0
Faker==6.1.1
flake8==3.8.4
future==0.18.2
gtin==0.1.13
idna==2.10
importlib-metadata==3.1.1
ipdb==0.13.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==3.2.0
mccabe==0.6.1
oauthlib==3.1.0
parameterized==0.8.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.8.0
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.3.0
rq==1.7.0
s3transfer==0.3.4
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
text-unidecode==1.3
traitlets==4.3.3
unicode-slugify==0.1.3
Unidecode==1.2.0
urllib3==1.25.11
waitress==1.4.4
wcwidth==0.2.5
WebOb==1.8.6
WebTest==2.0.35
Werkzeug==1.0.1
zipp==3.4.0
